Lecture_03_Langchain_Models_code_demo/requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
# langchain-openai
# openai

# Anthropic Integration
# langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
